requirements.txt
streamlit
pandas
scikit-learn
Pillow
